requirements.txt
sentence-transformers
faiss-cpu
langchain
langchain-groq
pydantic
python-dotenv